# docs/swagger.yaml
definitions:
  models.Diary:
    properties:
      content:
        example: This is content of the diary
        type: string
      date:
        example: "2021-01-12"
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error_message:
        example: invalid JSON format
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /diary/{year}/{quarter}:
    get:
      consumes:
      - application/json
      description: Get Content of the Diary
      parameters:
      - description: Year
        in: path
        name: year
        required: true
        type: integer
      - description: Quarter
        in: path
        name: quarter
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Diary'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Diary services
      tags:
      - diary
...
securityDefinitions:
  bearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"